Upgrade to latest Pug4J 2.0.0
