+++ .circleci/config.yml
@@ -1,12 +1,12 @@
 version: 2
 jobs:
   build:
-    working_directory: ~/spring-boot-starter-jade4j
+    working_directory: ~/spring-boot-starter-pug4j
     docker:
       - image: circleci/openjdk:8-jdk-browsers
     steps:
       - checkout
       - run: ./gradlew test jacocoTestReport coveralls
       # - save_cache:
-      #     key: spring-boot-starter-jade4j-{{ checksum "build.gradle" }}
+      #     key: spring-boot-starter-pug4j-{{ checksum "build.gradle" }}
