+++ gradle.properties
@@ -3,7 +3,7 @@ group=com.domingosuarez.boot
 version=2.0.5.RELEASE
 
 description=This Starter provides you the minimal and required configuration to use jade templates as views in your Spring Boot application.
-pomName=jade4j Spring Boot Starter
+pomName=pug4j Spring Boot Starter
 
 javaVersion=1.8
 springBootVersion=2.1.6.RELEASE
