+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'jade4j-spring-boot-starter'
+rootProject.name = 'pug4j-spring-boot-starter'
