+++ src/main/java/com/domingosuarez/boot/autoconfigure/pug4j/PugHelper.java
@@ -201,16 +201,34 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package com.domingosuarez.boot.autoconfigure.jade4j.support;
+package com.domingosuarez.boot.autoconfigure.pug4j;
 
-import com.domingosuarez.boot.autoconfigure.jade4j.JadeHelper;
+import org.springframework.core.annotation.AliasFor;
+import org.springframework.stereotype.Component;
+
+import java.lang.annotation.*;
 
 /**
- * Created by domix on 12/26/14.
+ * Indicates that an annotated class is a "Pug Helper" (e.g. utility object in the pug context).
+ * <p/>
+ * <p>This annotation serves as a specialization of {@link Component @Component},
+ * allowing for implementation classes to be autodetected through classpath scanning.
+ *
+ * @author Domingo Suarez Torres
+ * @see org.springframework.stereotype.Component
+ * @see org.springframework.context.annotation.ClassPathBeanDefinitionScanner
  */
-@JadeHelper
-public class SomeJadeUtil {
-  public String foo() {
-    return "foo";
-  }
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Component
+public @interface PugHelper {
+  /**
+   * The value may indicate a suggestion for a logical component name,
+   * to be turned into a Spring bean in case of an autodetected component.
+   *
+   * @return the suggested component name, if any
+   */
+  @AliasFor(annotation = Component.class, attribute = "value")
+  String value() default "";
 }
