+++ src/main/resources/META-INF/spring.factories
@@ -1,7 +1,7 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.domingosuarez.boot.autoconfigure.jade4j.Jade4JAutoConfiguration
+com.domingosuarez.boot.autoconfigure.pug4j.Pug4JAutoConfiguration
 
 # Template availability providers
 org.springframework.boot.autoconfigure.template.TemplateAvailabilityProvider=\
-com.domingosuarez.boot.autoconfigure.jade4j.Jade4JTemplateAvailabilityProvider
+com.domingosuarez.boot.autoconfigure.pug4j.Pug4JTemplateAvailabilityProvider
