+++ src/test/resources/templates/demo.pug
@@ -1,7 +1,7 @@
 html
   head
-    title Jade
+    title Pug
   body
-    h1 Jade - Template engine
+    h1 Pug - Template engine
     if user
       h2 With user
