--- src/test/resources/templates/demo_withHelper.jade
@@ -1,7 +1,0 @@
-html
-    head
-        title Jade
-    body
-        h1 Jade - Template engine
-        if user
-            h2 With user #{someJadeUtil.foo()}
