+++ src/test/resources/templates/demo_withHelper.pug
@@ -1,0 +1,7 @@
+html
+    head
+        title Pug
+    body
+        h1 Pug - Template engine
+        if user
+            h2 With user #{somePugUtil.foo()}
